fix: refactore monolith BackupService.cs and GrowthTimelineService.cs
