+++ Jellyfin.Plugin.JellyfinHelper.Tests/Services/Timeline/GrowthTimelinePerformanceTests.cs
@@ -1,4 +1,4 @@
-using System.Diagnostics;
+﻿using System.Diagnostics;
 using Jellyfin.Plugin.JellyfinHelper.Services.Timeline;
 using Xunit;
 using Xunit.Abstractions;
@@ -34,14 +34,14 @@ public class GrowthTimelinePerformanceTests(ITestOutputHelper output)
         }
 
         entries.Sort((a, b) => a.CreatedUtc.CompareTo(b.CreatedUtc));
-        var granularity = GrowthTimelineService.DetermineGranularity(earliest, now);
+        var granularity = TimelineAggregator.DetermineGranularity(earliest, now);
         // Act
         var sw = Stopwatch.StartNew();
-        var result = GrowthTimelineService.BuildCumulativeTimeline(entries, earliest, now, granularity);
+        var result = TimelineAggregator.BuildCumulativeTimeline(entries, earliest, now, granularity);
         sw.Stop();
 
         // Assert
-        output.WriteLine($"BuildCumulativeTimeline: {entries.Count} entries → {result.Count} points in {sw.ElapsedMilliseconds}ms (granularity: {granularity})");
+        output.WriteLine($"BuildCumulativeTimeline: {entries.Count} entries â†’ {result.Count} points in {sw.ElapsedMilliseconds}ms (granularity: {granularity})");
         Assert.True(sw.ElapsedMilliseconds < 2000, $"Took {sw.ElapsedMilliseconds}ms, expected < 2000ms");
         Assert.NotEmpty(result);
     }
@@ -70,11 +70,11 @@ public class GrowthTimelinePerformanceTests(ITestOutputHelper output)
 
         // Act
         var sw = Stopwatch.StartNew();
-        var result = GrowthTimelineService.BuildCumulativeTimeline(entries, earliest, now, "daily");
+        var result = TimelineAggregator.BuildCumulativeTimeline(entries, earliest, now, "daily");
         sw.Stop();
 
         // Assert
-        output.WriteLine($"BuildCumulativeTimeline (daily): {entries.Count} entries → {result.Count} points in {sw.ElapsedMilliseconds}ms");
+        output.WriteLine($"BuildCumulativeTimeline (daily): {entries.Count} entries â†’ {result.Count} points in {sw.ElapsedMilliseconds}ms");
         Assert.True(sw.ElapsedMilliseconds < 3000, $"Took {sw.ElapsedMilliseconds}ms, expected < 3000ms");
         Assert.NotEmpty(result);
     }
@@ -99,7 +99,7 @@ public class GrowthTimelinePerformanceTests(ITestOutputHelper output)
 
         // Act
         var sw = Stopwatch.StartNew();
-        var result = GrowthTimelineService.MergeSnapshotIntoTimeline(
+        var result = TimelineAggregator.MergeSnapshotIntoTimeline(
             existingPoints,
             now,
             5_000_000_000_000L,
@@ -108,7 +108,7 @@ public class GrowthTimelinePerformanceTests(ITestOutputHelper output)
         sw.Stop();
 
         // Assert
-        output.WriteLine($"MergeSnapshotIntoTimeline: {existingPoints.Count} existing → {result.Count} points in {sw.ElapsedMilliseconds}ms");
+        output.WriteLine($"MergeSnapshotIntoTimeline: {existingPoints.Count} existing â†’ {result.Count} points in {sw.ElapsedMilliseconds}ms");
         Assert.True(sw.ElapsedMilliseconds < 500, $"Took {sw.ElapsedMilliseconds}ms, expected < 500ms");
         Assert.NotEmpty(result);
     }
@@ -151,11 +151,11 @@ public class GrowthTimelinePerformanceTests(ITestOutputHelper output)
 
         // Act
         var sw = Stopwatch.StartNew();
-        var result = GrowthTimelineService.BuildIncrementalEntries(currentDirs, baseline, now);
+        var result = TimelineAggregator.BuildIncrementalEntries(currentDirs, baseline, now);
         sw.Stop();
 
         // Assert
-        output.WriteLine($"BuildIncrementalEntries: {currentDirs.Count} dirs, {baseline.Directories.Count} baseline → {result.Count} entries in {sw.ElapsedMilliseconds}ms");
+        output.WriteLine($"BuildIncrementalEntries: {currentDirs.Count} dirs, {baseline.Directories.Count} baseline â†’ {result.Count} entries in {sw.ElapsedMilliseconds}ms");
         Assert.True(sw.ElapsedMilliseconds < 1000, $"Took {sw.ElapsedMilliseconds}ms, expected < 1000ms");
         Assert.NotEmpty(result);
     }
@@ -180,11 +180,11 @@ public class GrowthTimelinePerformanceTests(ITestOutputHelper output)
 
         // Act
         var sw = Stopwatch.StartNew();
-        var result = GrowthTimelineService.ConsolidateToGranularity(points, "monthly");
+        var result = TimelineAggregator.ConsolidateToGranularity(points, "monthly");
         sw.Stop();
 
         // Assert
-        output.WriteLine($"ConsolidateToGranularity: {points.Count} daily → {result.Count} monthly in {sw.ElapsedMilliseconds}ms");
+        output.WriteLine($"ConsolidateToGranularity: {points.Count} daily â†’ {result.Count} monthly in {sw.ElapsedMilliseconds}ms");
         Assert.True(sw.ElapsedMilliseconds < 500, $"Took {sw.ElapsedMilliseconds}ms, expected < 500ms");
         Assert.True(result.Count < points.Count, "Consolidation should reduce point count");
     }
@@ -211,11 +211,11 @@ public class GrowthTimelinePerformanceTests(ITestOutputHelper output)
 
         // Act
         var sw = Stopwatch.StartNew();
-        var result = GrowthTimelineService.DeduplicateConsecutivePoints(points);
+        var result = TimelineAggregator.DeduplicateConsecutivePoints(points);
         sw.Stop();
 
         // Assert
-        output.WriteLine($"DeduplicateConsecutivePoints: {points.Count} → {result.Count} points in {sw.ElapsedMilliseconds}ms");
+        output.WriteLine($"DeduplicateConsecutivePoints: {points.Count} â†’ {result.Count} points in {sw.ElapsedMilliseconds}ms");
         Assert.True(sw.ElapsedMilliseconds < 200, $"Took {sw.ElapsedMilliseconds}ms, expected < 200ms");
         Assert.True(result.Count < points.Count, "Deduplication should reduce point count");
     }
@@ -240,11 +240,11 @@ public class GrowthTimelinePerformanceTests(ITestOutputHelper output)
 
         // Act
         var sw = Stopwatch.StartNew();
-        var result = GrowthTimelineService.TrimLeadingZeros(points);
+        var result = TimelineAggregator.TrimLeadingZeros(points);
         sw.Stop();
 
         // Assert
-        output.WriteLine($"TrimLeadingZeros: {points.Count} → {result.Count} points in {sw.ElapsedMilliseconds}ms");
+        output.WriteLine($"TrimLeadingZeros: {points.Count} â†’ {result.Count} points in {sw.ElapsedMilliseconds}ms");
         Assert.True(sw.ElapsedMilliseconds < 100, $"Took {sw.ElapsedMilliseconds}ms, expected < 100ms");
         Assert.True(result.Count < points.Count, "Trimming should reduce point count");
     }
@@ -284,11 +284,11 @@ public class GrowthTimelinePerformanceTests(ITestOutputHelper output)
 
         // Act
         var sw = Stopwatch.StartNew();
-        GrowthTimelineService.UpdateBaseline(baseline, currentDirs);
+        TimelineAggregator.UpdateBaseline(baseline, currentDirs);
         sw.Stop();
 
         // Assert
-        output.WriteLine($"UpdateBaseline: 8,000 baseline + 10,000 current (6,000 overlap, 4,000 new, 2,000 removed) → {baseline.Directories.Count} dirs in {sw.ElapsedMilliseconds}ms");
+        output.WriteLine($"UpdateBaseline: 8,000 baseline + 10,000 current (6,000 overlap, 4,000 new, 2,000 removed) â†’ {baseline.Directories.Count} dirs in {sw.ElapsedMilliseconds}ms");
         Assert.True(sw.ElapsedMilliseconds < 1000, $"Took {sw.ElapsedMilliseconds}ms, expected < 1000ms");
         Assert.Equal(10_000, baseline.Directories.Count);
