+++ Jellyfin.Plugin.JellyfinHelper/Services/Timeline/GrowthTimelineService.cs
@@ -20,6 +20,7 @@ namespace Jellyfin.Plugin.JellyfinHelper.Services.Timeline;
 ///     growth tracking on subsequent scans.
 ///     Automatically selects the best granularity (daily/weekly/monthly/quarterly/yearly)
 ///     depending on the time span between the oldest file and today.
+///     Pure aggregation logic is delegated to <see cref="TimelineAggregator" />.
 /// </summary>
 public sealed class GrowthTimelineService : IGrowthTimelineService, IDisposable
 {
@@ -108,8 +109,8 @@ public sealed class GrowthTimelineService : IGrowthTimelineService, IDisposable
             }
 
             var earliestExisting = existingTimeline.DataPoints[0].Date;
-            var granularity = DetermineGranularity(earliestExisting, now);
-            var zeroPoints = MergeSnapshotIntoTimeline(
+            var granularity = TimelineAggregator.DetermineGranularity(earliestExisting, now);
+            var zeroPoints = TimelineAggregator.MergeSnapshotIntoTimeline(
                 existingTimeline.DataPoints.ToList(),
                 now,
                 0,
@@ -117,9 +118,9 @@ public sealed class GrowthTimelineService : IGrowthTimelineService, IDisposable
                 granularity);
 
             // Run through the same finalization path as normal scans
-            zeroPoints = TrimLeadingZeros(zeroPoints);
-            zeroPoints = ConsolidateToGranularity(zeroPoints, granularity);
-            zeroPoints = DeduplicateConsecutivePoints(zeroPoints);
+            zeroPoints = TimelineAggregator.TrimLeadingZeros(zeroPoints);
+            zeroPoints = TimelineAggregator.ConsolidateToGranularity(zeroPoints, granularity);
+            zeroPoints = TimelineAggregator.DeduplicateConsecutivePoints(zeroPoints);
 
             var zeroResult = new GrowthTimelineResult
             {
@@ -192,14 +193,14 @@ public sealed class GrowthTimelineService : IGrowthTimelineService, IDisposable
             timelineEntries.Sort((a, b) => a.CreatedUtc.CompareTo(b.CreatedUtc));
 
             var earliest = timelineEntries.Count > 0 ? timelineEntries[0].CreatedUtc : now;
-            var granularity = DetermineGranularity(earliest, now);
+            var granularity = TimelineAggregator.DetermineGranularity(earliest, now);
 
             _pluginLog.LogInfo(
                 "GrowthTimeline",
                 $"Building initial timeline: {timelineEntries.Count} entries, earliest: {earliest:yyyy-MM-dd}, granularity: {granularity}",
                 _logger);
 
-            dataPoints = BuildCumulativeTimeline(timelineEntries, earliest, now, granularity);
+            dataPoints = TimelineAggregator.BuildCumulativeTimeline(timelineEntries, earliest, now, granularity);
         }
         else
         {
@@ -229,9 +230,9 @@ public sealed class GrowthTimelineService : IGrowthTimelineService, IDisposable
                     _logger);
 
                 var earliestExisting = existingTimeline.DataPoints[0].Date;
-                var granularity = DetermineGranularity(earliestExisting, now);
+                var granularity = TimelineAggregator.DetermineGranularity(earliestExisting, now);
 
-                dataPoints = MergeSnapshotIntoTimeline(
+                dataPoints = TimelineAggregator.MergeSnapshotIntoTimeline(
                     existingTimeline.DataPoints.ToList(),
                     now,
                     currentTotalSize,
@@ -247,17 +248,17 @@ public sealed class GrowthTimelineService : IGrowthTimelineService, IDisposable
                     "No existing timeline found. Performing historical reconstruction from baseline.",
                     _logger);
 
-                var timelineEntries = BuildIncrementalEntries(currentDirs, baseline, now);
+                var timelineEntries = TimelineAggregator.BuildIncrementalEntries(currentDirs, baseline, now);
                 timelineEntries.Sort((a, b) => a.CreatedUtc.CompareTo(b.CreatedUtc));
 
                 var earliest = timelineEntries.Count > 0 ? timelineEntries[0].CreatedUtc : now;
-                var granularity = DetermineGranularity(earliest, now);
+                var granularity = TimelineAggregator.DetermineGranularity(earliest, now);
 
-                dataPoints = BuildCumulativeTimeline(timelineEntries, earliest, now, granularity);
+                dataPoints = TimelineAggregator.BuildCumulativeTimeline(timelineEntries, earliest, now, granularity);
             }
 
             // Update baseline with current state for next scan
-            UpdateBaseline(baseline, currentDirs);
+            TimelineAggregator.UpdateBaseline(baseline, currentDirs);
             await SaveBaselineAsync(baseline, cancellationToken).ConfigureAwait(false);
         }
 
@@ -266,20 +267,20 @@ public sealed class GrowthTimelineService : IGrowthTimelineService, IDisposable
         // Trim leading zero-value data points but keep one zero just before the first non-zero
         // as a visual baseline start. This avoids long flat 0-lines for historical buckets
         // before any media existed, while still showing a library rebuild (drop to 0 then rise).
-        dataPoints = TrimLeadingZeros(dataPoints);
+        dataPoints = TimelineAggregator.TrimLeadingZeros(dataPoints);
 
         // Consolidate data points into the current granularity.
         // When the time span grows (e.g. from <90 days to >90 days), the granularity
         // upgrades (daily→weekly). Previously stored finer-grained points are merged
         // into the coarser buckets so the persisted file stays compact.
         var finalGranularity = dataPoints.Count > 0
-            ? DetermineGranularity(dataPoints[0].Date, now)
+            ? TimelineAggregator.DetermineGranularity(dataPoints[0].Date, now)
             : "monthly";
-        dataPoints = ConsolidateToGranularity(dataPoints, finalGranularity);
+        dataPoints = TimelineAggregator.ConsolidateToGranularity(dataPoints, finalGranularity);
 
         // Remove consecutive data points with identical values to reduce storage size.
         // The UI will interpolate missing buckets back when rendering the chart.
-        dataPoints = DeduplicateConsecutivePoints(dataPoints);
+        dataPoints = TimelineAggregator.DeduplicateConsecutivePoints(dataPoints);
 
         if (dataPoints.Count == 0)
         {
@@ -351,173 +352,6 @@ public sealed class GrowthTimelineService : IGrowthTimelineService, IDisposable
         finally
         {
             _fileLock.Release();
-        }
-    }
-
-    /// <summary>
-    ///     Determines the best granularity based on the time span between the oldest file and now.
-    /// </summary>
-    /// <param name="earliest">The earliest file date.</param>
-    /// <param name="now">The current date.</param>
-    /// <returns>The granularity string (daily, weekly, monthly, quarterly, yearly).</returns>
-    internal static string DetermineGranularity(DateTime earliest, DateTime now)
-    {
-        var span = now - earliest;
-        var totalDays = span.TotalDays;
-
-        return totalDays switch
-        {
-            > 5 * 365 => "yearly",
-            > 2 * 365 => "quarterly",
-            > 365 => "monthly",
-            > 90 => "weekly",
-            _ => "daily"
-        };
-    }
-
-    /// <summary>
-    ///     Generates bucket start dates from earliest to now based on granularity.
-    /// </summary>
-    /// <param name="earliest">The earliest date to start from.</param>
-    /// <param name="now">The current date (upper bound).</param>
-    /// <param name="granularity">The granularity (daily, weekly, monthly, quarterly, yearly).</param>
-    /// <returns>A list of bucket start dates.</returns>
-    internal static List<DateTime> GenerateBucketStarts(DateTime earliest, DateTime now, string granularity)
-    {
-        var buckets = new List<DateTime>();
-        var current = GetBucketStart(earliest, granularity);
-
-        while (current <= now)
-        {
-            buckets.Add(current);
-            current = AdvanceBucket(current, granularity);
-        }
-
-        return buckets;
-    }
-
-    /// <summary>
-    ///     Builds incremental timeline entries by comparing the current directory state
-    ///     against the baseline. Baseline directories contribute their full size at their
-    ///     creation date. New directories (created after firstScanTimestamp) also contribute
-    ///     their full size at their creation date. Size changes in baseline directories
-    ///     contribute only the positive diff at the scan timestamp.
-    /// </summary>
-    /// <param name="currentDirs">The currently scanned directories.</param>
-    /// <param name="baseline">The baseline from the first scan.</param>
-    /// <param name="now">The current scan timestamp.</param>
-    /// <returns>A list of file entries for timeline construction.</returns>
-    internal static List<FileEntry> BuildIncrementalEntries(
-        List<DirectoryEntry> currentDirs,
-        GrowthTimelineBaseline baseline,
-        DateTime now)
-    {
-        var entries = (from kvp in baseline.Directories
-                let count = kvp.Value.Count > 0 ? kvp.Value.Count : 1
-                select new FileEntry { CreatedUtc = kvp.Value.CreatedUtc, Size = kvp.Value.Size, CountDelta = count })
-            .ToList();
-
-        // 1. Add all baseline entries at their original creation date with their original size
-
-        // 2. Process current directories
-        foreach (var dir in currentDirs)
-        {
-            if (baseline.Directories.TryGetValue(dir.Path, out var baselineEntry))
-            {
-                // Existing directory: check for size change
-                var sizeDiff = dir.Size - baselineEntry.Size;
-                switch (sizeDiff)
-                {
-                    case > 0:
-                    // Directory shrank: add a negative entry at the current scan time
-                    // CountDelta = 0 because this is a size adjustment, not a removal
-                    case < 0:
-                        // Directory grew: add the positive diff at the current scan time
-                        // CountDelta = 0 because this is a size adjustment, not a new directory
-                        entries.Add(
-                            new FileEntry
-                            {
-                                CreatedUtc = now,
-                                Size = sizeDiff,
-                                CountDelta = 0
-                            });
-                        break;
-                }
-
-                // No change: no entry needed (baseline already covers it)
-            }
-            else
-            {
-                // New group (not in baseline)
-                var count = dir.Count > 0 ? dir.Count : 1;
-                if (dir.CreatedUtc > baseline.FirstScanTimestamp)
-                {
-                    // Created after baseline: add full size at creation date
-                }
-
-                // Created before baseline but wasn't in the baseline
-                // (e.g. a new library location was added). Treat as baseline-era entry.
-                entries.Add(
-                    new FileEntry
-                    {
-                        CreatedUtc = dir.CreatedUtc,
-                        Size = dir.Size,
-                        CountDelta = count
-                    });
-            }
-        }
-
-        // 3. Handle deleted directories (in baseline but not in current scan)
-        var currentPaths = new HashSet<string>(currentDirs.Select(d => d.Path), StringComparer.OrdinalIgnoreCase);
-        entries.AddRange(
-            from kvp in baseline.Directories
-            where !currentPaths.Contains(kvp.Key)
-            let removedCount = kvp.Value.Count > 0 ? kvp.Value.Count : 1
-            select new FileEntry { CreatedUtc = now, Size = -kvp.Value.Size, CountDelta = -removedCount });
-
-        return entries;
-    }
-
-    /// <summary>
-    ///     Updates the baseline with the current directory state for subsequent scans.
-    ///     New directories are added to the baseline. Existing entries are updated with
-    ///     the current size. Removed directories are kept in the baseline (their deletion
-    ///     is tracked via negative diff entries).
-    /// </summary>
-    /// <param name="baseline">The baseline to update.</param>
-    /// <param name="currentDirs">The current directory entries.</param>
-    internal static void UpdateBaseline(GrowthTimelineBaseline baseline, List<DirectoryEntry> currentDirs)
-    {
-        var currentPaths = new HashSet<string>(StringComparer.OrdinalIgnoreCase);
-
-        foreach (var dir in currentDirs)
-        {
-            currentPaths.Add(dir.Path);
-
-            if (baseline.Directories.TryGetValue(dir.Path, out var existing))
-            {
-                // Update size and count to current values (creation date stays the same)
-                existing.Size = dir.Size;
-                existing.Count = dir.Count;
-            }
-            else
-            {
-                // New group: add to baseline
-                baseline.Directories[dir.Path] = new BaselineDirectoryEntry
-                {
-                    CreatedUtc = dir.CreatedUtc,
-                    Size = dir.Size,
-                    Count = dir.Count
-                };
-            }
-        }
-
-        // Remove deleted directories from baseline so they don't generate
-        // spurious negative diffs on future scans
-        var toRemove = baseline.Directories.Keys.Where(k => !currentPaths.Contains(k)).ToList();
-        foreach (var key in toRemove)
-        {
-            baseline.Directories.Remove(key);
         }
     }
 
@@ -670,142 +504,6 @@ public sealed class GrowthTimelineService : IGrowthTimelineService, IDisposable
     }
 
     /// <summary>
-    ///     Builds cumulative data points from sorted file entries using the specified granularity.
-    /// </summary>
-    /// <param name="sortedEntries">The file entries sorted by creation date.</param>
-    /// <param name="earliest">The earliest date to start from.</param>
-    /// <param name="now">The current date (upper bound).</param>
-    /// <param name="granularity">The granularity (daily, weekly, monthly, quarterly, yearly).</param>
-    /// <returns>A list of cumulative growth timeline data points.</returns>
-    internal static List<GrowthTimelinePoint> BuildCumulativeTimeline(
-        List<FileEntry> sortedEntries,
-        DateTime earliest,
-        DateTime now,
-        string granularity)
-    {
-        var points = new List<GrowthTimelinePoint>();
-
-        // Generate all bucket start dates
-        var bucketStarts = GenerateBucketStarts(earliest, now, granularity);
-        if (bucketStarts.Count == 0)
-        {
-            return points;
-        }
-
-        var fileIndex = 0;
-        long cumulativeSize = 0;
-        var cumulativeCount = 0;
-
-        for (var b = 0; b < bucketStarts.Count; b++)
-        {
-            var bucketStart = bucketStarts[b];
-            var bucketEnd = b + 1 < bucketStarts.Count ? bucketStarts[b + 1] : now.AddDays(1);
-
-            // Accumulate all files whose creation date falls before the bucket end
-            while (fileIndex < sortedEntries.Count && sortedEntries[fileIndex].CreatedUtc < bucketEnd)
-            {
-                cumulativeSize += sortedEntries[fileIndex].Size;
-                cumulativeCount += sortedEntries[fileIndex].CountDelta;
-                fileIndex++;
-            }
-
-            points.Add(
-                new GrowthTimelinePoint
-                {
-                    Date = bucketStart,
-                    CumulativeSize = cumulativeSize,
-                    CumulativeFileCount = cumulativeCount
-                });
-        }
-
-        return points;
-    }
-
-    /// <summary>
-    ///     Merges a current snapshot into an existing timeline using append-only semantics.
-    ///     All existing data points whose bucket date is strictly before the current bucket
-    ///     are preserved as immutable history. The current bucket is replaced (or added)
-    ///     with the actual current total size and count. This ensures that deletions of
-    ///     files with past creation dates do not retroactively alter historical data points;
-    ///     instead, the deletion manifests as a size drop at the current point in time.
-    /// </summary>
-    /// <param name="existingPoints">The previously persisted data points (chronologically sorted).</param>
-    /// <param name="now">The current scan timestamp.</param>
-    /// <param name="currentTotalSize">The absolute total size of all current media directories.</param>
-    /// <param name="currentTotalCount">The absolute total count of all current media items.</param>
-    /// <param name="granularity">The granularity for bucket calculation.</param>
-    /// <returns>A merged list of data points with the current snapshot appended/updated.</returns>
-    internal static List<GrowthTimelinePoint> MergeSnapshotIntoTimeline(
-        List<GrowthTimelinePoint> existingPoints,
-        DateTime now,
-        long currentTotalSize,
-        int currentTotalCount,
-        string granularity)
-    {
-        var currentBucketStart = GetBucketStart(now, granularity);
-
-        // Keep all points strictly before the current bucket (immutable history)
-        var result = existingPoints
-            .Where(point => point.Date < currentBucketStart)
-            .ToList();
-
-        // Add the current snapshot as the latest data point
-        result.Add(
-            new GrowthTimelinePoint
-            {
-                Date = currentBucketStart,
-                CumulativeSize = currentTotalSize,
-                CumulativeFileCount = currentTotalCount
-            });
-
-        return result;
-    }
-
-    /// <summary>
-    ///     Gets the start of the bucket containing the given date.
-    /// </summary>
-    /// <param name="date">The date to find the bucket start for.</param>
-    /// <param name="granularity">The granularity (daily, weekly, monthly, quarterly, yearly).</param>
-    /// <returns>The start date of the bucket containing the given date.</returns>
-    internal static DateTime GetBucketStart(DateTime date, string granularity)
-    {
-        return granularity switch
-        {
-            "daily" => new DateTime(date.Year, date.Month, date.Day, 0, 0, 0, DateTimeKind.Utc),
-            "weekly" => GetStartOfWeek(date),
-            "monthly" => new DateTime(date.Year, date.Month, 1, 0, 0, 0, DateTimeKind.Utc),
-            "quarterly" => new DateTime(date.Year, ((date.Month - 1) / 3 * 3) + 1, 1, 0, 0, 0, DateTimeKind.Utc),
-            "yearly" => new DateTime(date.Year, 1, 1, 0, 0, 0, DateTimeKind.Utc),
-            _ => new DateTime(date.Year, date.Month, 1, 0, 0, 0, DateTimeKind.Utc)
-        };
-    }
-
-    /// <summary>
-    ///     Advances a bucket start to the next bucket.
-    /// </summary>
-    private static DateTime AdvanceBucket(DateTime current, string granularity)
-    {
-        return granularity switch
-        {
-            "daily" => current.AddDays(1),
-            "weekly" => current.AddDays(7),
-            "monthly" => current.AddMonths(1),
-            "quarterly" => current.AddMonths(3),
-            "yearly" => current.AddYears(1),
-            _ => current.AddMonths(1)
-        };
-    }
-
-    /// <summary>
-    ///     Gets the start of the ISO week (Monday) for a given date.
-    /// </summary>
-    private static DateTime GetStartOfWeek(DateTime date)
-    {
-        var diff = ((int)date.DayOfWeek - (int)DayOfWeek.Monday + 7) % 7;
-        return new DateTime(date.Year, date.Month, date.Day, 0, 0, 0, DateTimeKind.Utc).AddDays(-diff);
-    }
-
-    /// <summary>
     ///     Loads the baseline from disk.
     /// </summary>
     /// <param name="cancellationToken">A token to monitor for cancellation requests.</param>
@@ -902,126 +600,6 @@ public sealed class GrowthTimelineService : IGrowthTimelineService, IDisposable
         {
             _fileLock.Release();
         }
-    }
-
-    /// <summary>
-    ///     Removes leading zero-value data points from the timeline, keeping at most one
-    ///     zero point immediately before the first non-zero point as a visual baseline start.
-    ///     Only trims from the beginning — zero points in the middle or end of the timeline
-    ///     (e.g. after a library deletion) are preserved to show the full history.
-    /// </summary>
-    /// <param name="points">The data points to trim.</param>
-    /// <returns>A trimmed list with leading zeros removed.</returns>
-    internal static List<GrowthTimelinePoint> TrimLeadingZeros(List<GrowthTimelinePoint> points)
-    {
-        if (points.Count == 0)
-        {
-            return points;
-        }
-
-        // Find the index of the first non-zero data point
-        var firstNonZero = -1;
-        for (var i = 0; i < points.Count; i++)
-        {
-            if (points[i].CumulativeSize <= 0 && points[i].CumulativeFileCount <= 0)
-            {
-                continue;
-            }
-
-            firstNonZero = i;
-            break;
-        }
-
-        switch (firstNonZero)
-        {
-            // All points are zero — return empty (nothing meaningful to show)
-            case < 0:
-                return [];
-            // No leading zeros — return as-is
-            case 0:
-                return points;
-            default:
-            {
-                // Keep one zero point before the first non-zero as the visual "start from zero" baseline
-                var startIndex = firstNonZero - 1;
-                return points.GetRange(startIndex, points.Count - startIndex);
-            }
-        }
-    }
-
-    /// <summary>
-    ///     Consolidates data points from a finer granularity into a coarser one.
-    ///     When the time span grows and the granularity upgrades (e.g. daily→weekly),
-    ///     multiple finer-grained points that fall into the same coarser bucket are merged
-    ///     by keeping the last (most recent) point per bucket. This keeps the persisted
-    ///     timeline compact and ensures backup/restore works with consolidated data.
-    /// </summary>
-    /// <param name="points">The data points (sorted chronologically).</param>
-    /// <param name="targetGranularity">The target granularity to consolidate into.</param>
-    /// <returns>A consolidated list with at most one point per target bucket.</returns>
-    internal static List<GrowthTimelinePoint> ConsolidateToGranularity(
-        List<GrowthTimelinePoint> points,
-        string targetGranularity)
-    {
-        if (points.Count <= 1)
-        {
-            return points;
-        }
-
-        // Group points by their target bucket start date.
-        // For each bucket, keep the last point (highest cumulative values).
-        var buckets = new Dictionary<DateTime, GrowthTimelinePoint>();
-
-        foreach (var point in points)
-        {
-            var bucketStart = GetBucketStart(point.Date, targetGranularity);
-            // Overwrite: the last point per bucket wins (points are sorted chronologically)
-            buckets[bucketStart] = new GrowthTimelinePoint
-            {
-                Date = bucketStart,
-                CumulativeSize = point.CumulativeSize,
-                CumulativeFileCount = point.CumulativeFileCount
-            };
-        }
-
-        var result = buckets.Values.OrderBy(p => p.Date).ToList();
-        return result;
-    }
-
-    /// <summary>
-    ///     Removes consecutive data points that have identical CumulativeSize and CumulativeFileCount.
-    ///     Only the first point of each "plateau" is kept, plus the last point is always preserved
-    ///     so the timeline's time span remains correct. The UI is responsible for interpolating
-    ///     the missing intermediate buckets when rendering the chart.
-    /// </summary>
-    /// <param name="points">The data points (already sorted chronologically).</param>
-    /// <returns>A deduplicated list with redundant consecutive points removed.</returns>
-    internal static List<GrowthTimelinePoint> DeduplicateConsecutivePoints(List<GrowthTimelinePoint> points)
-    {
-        if (points.Count <= 2)
-        {
-            return points;
-        }
-
-        var result = new List<GrowthTimelinePoint> { points[0] };
-
-        for (var i = 1; i < points.Count - 1; i++)
-        {
-            var prev = points[i - 1];
-            var curr = points[i];
-
-            // Keep the point if it differs from its predecessor
-            if (curr.CumulativeSize != prev.CumulativeSize ||
-                curr.CumulativeFileCount != prev.CumulativeFileCount)
-            {
-                result.Add(curr);
-            }
-        }
-
-        // Always keep the last point to preserve the timeline's end date
-        result.Add(points[^1]);
-
-        return result;
     }
 
     /// <summary>
